refactor: BottomNavigation에 사용되는 중복 로직 추출 (#10)
