+++ src/components/icons/SelectIcon.tsx
@@ -1,21 +1,9 @@
-import type { ComponentProps } from 'react';
-
-// Todo: NavIcon을 만들어서 중복되는 로직 분리하기!
-export const navIconColors = {
-  default: '#BDBDBD',
-  selected: '#FF9081',
-} as const;
-
-export type NavIconProps = ComponentProps<'svg'> & {
-  color: keyof typeof navIconColors;
-};
+import type { NavIconProps } from '@/components/layout/BottomNavigation';
 
 export default function SelecIcon({
-  color = 'default',
+  color = '#BDBDBD',
   ...props
 }: NavIconProps) {
-  const svgColor = navIconColors[color];
-
   return (
     <svg
       xmlns="http://www.w3.org/2000/svg"
@@ -26,8 +14,8 @@ export default function SelecIcon({
       {...props}
     >
       <path
-        fill={svgColor}
-        fill-rule="evenodd"
+        fill={color}
+        fillRule="evenodd"
         d="m25.666 11.439 13.918 5.128c1.925.71 3.565 1.312 4.755 1.962 1.245.683 2.425 1.64 2.78 3.255s-.315 2.98-1.16 4.12c-.808 1.09-2.045 2.325-3.495 3.775l-1.785 1.785 8.563 8.56c.475.475.895.898 1.222 1.273.348.397.678.842.907 1.395a4.97 4.97 0 0 1 0 3.8c-.223.512-.53.984-.907 1.397-.328.375-.75.795-1.225 1.27l-.08.083a27.91 27.91 0 0 1-1.27 1.222 5.073 5.073 0 0 1-1.395.907 4.97 4.97 0 0 1-3.803 0 5.074 5.074 0 0 1-1.395-.907c-.374-.328-.797-.75-1.272-1.225l-8.56-8.56-1.785 1.785c-1.45 1.45-2.685 2.688-3.775 3.495-1.14.845-2.505 1.515-4.12 1.16-1.615-.355-2.575-1.535-3.255-2.78-.65-1.19-1.253-2.83-1.962-4.755l-5.128-13.918c-1.45-3.94-2.615-7.1-3.132-9.532-.515-2.428-.545-4.843 1.22-6.607 1.764-1.765 4.18-1.735 6.607-1.22 2.433.517 5.593 1.68 9.532 3.132Zm-10.31.537c-2.17-.462-2.842-.132-3.177.2-.335.338-.665 1.01-.203 3.18.45 2.135 1.515 5.038 3.04 9.17l5.018 13.625c.775 2.1 1.285 3.475 1.785 4.388.47.86.715.907.765.917h.01c.05.015.29.075 1.075-.51.84-.622 1.88-1.654 3.462-3.237l3.008-3.008a1.875 1.875 0 0 1 2.65 0l9.847 9.848c.525.525.858.855 1.125 1.087.108.108.231.199.365.27.3.126.633.126.933 0 .134-.071.257-.162.365-.27.265-.232.598-.56 1.125-1.087.525-.53.855-.86 1.087-1.125.108-.108.199-.23.27-.365a1.22 1.22 0 0 0 0-.932 1.566 1.566 0 0 0-.27-.366 25.996 25.996 0 0 0-1.087-1.125l-9.847-9.847a1.875 1.875 0 0 1 0-2.65l3.007-3.008c1.582-1.582 2.615-2.622 3.237-3.46.583-.787.526-1.027.513-1.077v-.005l-.003-.005c-.01-.05-.057-.295-.917-.765-.915-.5-2.288-1.01-4.388-1.785l-13.627-5.017c-4.133-1.525-7.033-2.588-9.168-3.04Z"
         clip-rule="evenodd"
       />
