+++ src/components/layout/BottomNavigation.tsx
@@ -1,9 +1,19 @@
+import { memo, type ComponentProps } from 'react';
 import Link from 'next/link';
 import { useRouter } from 'next/router';
+import { getNavItemStatus } from '@/util/route';
 import BoardIcon from '../icons/BoardIcon';
 import ChatRoomIcon from '../icons/ChatRoomIcon';
 import SelectIcon from '../icons/SelectIcon';
-import { isSelected } from '@/util/route';
+
+export const navIconColors = {
+  default: '#BDBDBD',
+  selected: '#FF9081',
+} as const;
+
+export type NavIconProps = ComponentProps<'svg'> & {
+  color: (typeof navIconColors)[keyof typeof navIconColors];
+};
 
 const NAV_ITEMS = [
   { href: '/chatroom', Icon: ChatRoomIcon, label: '대화상세' },
@@ -11,15 +21,26 @@ const NAV_ITEMS = [
   { href: '/status', Icon: BoardIcon, label: '질문현황' },
 ];
 
-export default function BottomNavigation() {
+type NavItmeProps = (typeof NAV_ITEMS)[number];
+
+const NavItem = memo(({ href, Icon, label }: NavItmeProps) => {
   const router = useRouter();
+  const color = navIconColors[getNavItemStatus(router.pathname, href)];
+
+  return (
+    <Link href={href}>
+      <Icon color={color} />
+      <span className="text-xs">{label}</span>
+    </Link>
+  );
+});
+NavItem.displayName = 'NavItem';
+
+export default function BottomNavigation() {
   return (
     <nav className="btm-nav btm-nav-lg max-w-[512px] mx-auto">
-      {NAV_ITEMS.map(({ href, Icon, label }) => (
-        <Link key={href} href={href}>
-          <Icon color={isSelected(router.pathname, href)} />
-          <span className="text-xs">{label}</span>
-        </Link>
+      {NAV_ITEMS.map((navItem) => (
+        <NavItem key={navItem.href} {...navItem} />
       ))}
     </nav>
   );
