+++ src/util/route.ts
@@ -1,3 +1,3 @@
-export function isSelected(pathname: string, href: string) {
+export function getNavItemStatus(pathname: string, href: string) {
   return new RegExp(href).test(pathname) ? 'selected' : 'default';
 }
